Use a Transformer model in the multi features example
